Scipy 1.6.1

+++ pyodide-build/pyodide_build/buildpkg.py
@@ -241,9 +241,10 @@ def prepare_source(
         The location where the source ended up.
     """
     if "url" in src_metadata:
-        srcpath = download_and_extract(buildpath, srcpath, src_metadata)
+        extract_dir = download_and_extract(buildpath, srcpath, src_metadata)
+        shutil.move(extract_dir, srcpath)
         patch(pkg_root, srcpath, src_metadata)
-        return srcpath
+        return
 
     if "path" not in src_metadata:
         raise ValueError(
@@ -257,8 +258,6 @@ def prepare_source(
 
     if not srcpath.is_dir():
         shutil.copytree(srcdir, srcpath)
-
-    return srcpath
 
 
 def patch(pkg_root: Path, srcpath: Path, src_metadata: Dict[str, Any]):
@@ -331,6 +330,7 @@ def compile(
     *,
     target_install_dir: str,
     host_install_dir: str,
+    continue_from: int = 0,
 ):
     """
     Runs pywasmcross for the package. The effect of this is to first run setup.py
@@ -376,11 +376,22 @@ def compile(
     replace_libs = ";".join(build_metadata.get("replace-libs", []))
 
     with chdir(srcpath):
-        pywasmcross.capture_compile(
-            host_install_dir=host_install_dir,
-            skip_host=skip_host,
-            env=bash_runner.env,
-        )
+        if not continue_from:
+            pywasmcross.capture_compile(
+                host_install_dir=host_install_dir,
+                skip_host=skip_host,
+                env=bash_runner.env,
+            )
+            if Path("../../prereplay_patches").exists():
+                for f in Path("../../prereplay_patches").glob("*.patch"):
+                    subprocess.run(
+                        [
+                            "patch",
+                            "-p1",
+                            "-i",
+                            str(f),
+                        ]
+                    )
         pywasmcross.replay_compile(
             cflags=build_metadata["cflags"],
             cxxflags=build_metadata["cxxflags"],
@@ -388,6 +399,7 @@ def compile(
             target_install_dir=target_install_dir,
             host_install_dir=host_install_dir,
             replace_libs=replace_libs,
+            continue_from=continue_from,
         )
         install_for_distribution()
 
@@ -441,6 +453,7 @@ def unvendor_tests(install_prefix: Path, test_install_prefix: Path) -> int:
         if root_rel.name in ["test", "tests"]:
             # This is a test folder
             (test_install_prefix / root_rel).parent.mkdir(exist_ok=True, parents=True)
+            (test_install_prefix / root_rel).unlink(missing_ok=True)
             shutil.move(install_prefix / root_rel, test_install_prefix / root_rel)
             n_moved += 1
             continue
@@ -613,6 +626,7 @@ def build_package(
     target_install_dir: str,
     host_install_dir: str,
     compress_package: bool,
+    continue_from: int = 0,
 ):
     """
     Build the package. The main entrypoint in this module.
@@ -636,23 +650,27 @@ def build_package(
     name = pkg["package"]["name"]
     build_dir = pkg_root / "build"
     src_dir_name: str = name + "-" + pkg["package"]["version"]
-    src_path = build_dir / src_dir_name
+    srcpath = build_dir / src_dir_name
     source_metadata = pkg["source"]
     build_metadata = pkg["build"]
     with chdir(pkg_root), get_bash_runner() as bash_runner:
         if not needs_rebuild(pkg_root, build_dir, source_metadata):
             return
-        if source_metadata:
-            if build_dir.resolve().is_dir():
-                shutil.rmtree(build_dir)
-            os.makedirs(build_dir)
+        if continue_from:
+            assert srcpath.exists()
+        else:
+            if source_metadata:
+                if build_dir.resolve().is_dir():
+                    shutil.rmtree(build_dir)
+                os.makedirs(build_dir)
 
-        srcpath = prepare_source(pkg_root, build_dir, src_path, source_metadata)
-        if build_metadata.get("script"):
-            run_script(build_dir, srcpath, build_metadata, bash_runner)
-        if build_metadata.get("library"):
-            create_packaged_token(build_dir)
-            return
+            prepare_source(pkg_root, build_dir, srcpath, source_metadata)
+            if build_metadata.get("script"):
+                run_script(build_dir, srcpath, build_metadata, bash_runner)
+            if build_metadata.get("library"):
+                create_packaged_token(build_dir)
+                return
+
         # shared libraries get built by the script and put into install
         # subfolder, then packaged into a pyodide module
         # i.e. they need package running, but not compile
@@ -664,6 +682,7 @@ def build_package(
                 bash_runner,
                 target_install_dir=target_install_dir,
                 host_install_dir=host_install_dir,
+                continue_from=continue_from,
             )
         should_unvendor_tests = build_metadata.get("unvendor-tests", True)
         package_files(
@@ -725,6 +744,12 @@ def make_parser(parser: argparse.ArgumentParser):
         ),
     )
     parser.add_argument(
+        "--continue-from",
+        type=int,
+        nargs="?",
+        default=0,
+    )
+    parser.add_argument(
         "--no-compress-package",
         action="store_false",
         default=True,
@@ -764,6 +789,7 @@ def main(args):
             target_install_dir=args.target_install_dir,
             host_install_dir=args.host_install_dir,
             compress_package=args.compress_package,
+            continue_from=args.continue_from,
         )
     except:
         success = False
